Move loggers to labor directory, switch perform to run, decrease redis hits collecting metrics
