+++ lib/tasks/fetch.rake
@@ -127,5 +127,5 @@ task record_db_table_counts: :environment do
 end
 
 task log_worker_queue_stats: :environment do
-  LogWorkerQueueStats.perform
+  Loggers::LogWorkerQueueStats.perform
 end
